Indent code and organize scripts in one file (#7114)
