+++ flow-dnd/src/main/java/com/vaadin/flow/component/dnd/DragSource.java
@@ -170,9 +170,7 @@ public interface DragSource<T extends Component> extends HasElement {
     /**
      * Sets this component as draggable. By default it is not.
      *
-     * @param draggable
-     *            {@code true} for enable dragging, {@code f
-     *            alse} to prevent
+     * @param draggable {@code true} for enable dragging, {@code false} to prevent
      */
     default void setDraggable(boolean draggable) {
         if (draggable == isDraggable()) {
